[#5] feat: Character struct 추가
CharacterType enum을 통해 각 캐릭터 타입마다 필요한 assets을 그룹핑 했다면 struct를 이용 같은 타입이라도 매일 갱신되는 서로 다른 캐릭터가 다른 프로퍼티를 가질 수 있도록 함
* init를 통해 현재 name의 경우는 일단 case 를 따르지만 향후 사용자에게 입력을 받아 각자 다른 값을 줄 수 있음
